Renamings. Move questions to separate file. Cleanup console.log. Make stats a function.
