+++ js/imageType.js
@@ -1,0 +1,6 @@
+const imageType = {
+  PHOTO: `photo`,
+  PAINT: `paint`
+};
+
+export default imageType;
